Improve saving functionality
This fixes a few issues with the saving and loading functionality.
Previously, when a file loaded, this plugin was pretty much only relying
on the functionality of the mkview function and was not actually loading
the position from the file. Now, I have more checks in place to make
sure we know ahead of time how the cursor needs to be placed while still
maintaining the file's fold state.